# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Minecraft Operator API
  description: |
    REST API for managing PaperMC servers and plugins in Kubernetes.

    This API provides CRUD operations for PaperMCServer and Plugin custom resources,
    as well as action endpoints for triggering reconciliation and immediate updates.

    ## Authentication

    The API uses the same authentication as the Kubernetes cluster (service account tokens).

    ## Error Handling

    All errors return a JSON object with `error` field containing human-readable message,
    optional `code` for programmatic handling, and optional `details` for additional context.
  version: v1
  contact:
    name: Aleksei Sviridkin
    email: f@lex.la
  license:
    name: BSD-3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /servers:
    get:
      summary: List all PaperMC servers
      description: |
        Returns a list of all PaperMCServer resources in the cluster.
        Can be filtered by namespace using query parameter.
      operationId: listServers
      tags:
        - servers
      parameters:
        - name: namespace
          in: query
          description: Filter servers by Kubernetes namespace
          required: false
          schema:
            type: string
            example: minecraft
      responses:
        "200":
          description: Successfully retrieved list of servers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerListResponse"
              example:
                servers:
                  - name: survival-server
                    namespace: minecraft
                    currentVersion: "1.21.1-100"
                    desiredVersion: "1.21.1-105"
                    updateStrategy: auto
                    status: running
                    pluginCount: 5
                    availableUpdate:
                      version: "1.21.1"
                      build: 105
                      releasedAt: "2025-01-15T10:30:00Z"
                    nextMaintenance: "Every Sunday at 4:00"
                    labels:
                      environment: production
                      game-mode: survival
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new PaperMC server
      description: |
        Creates a new PaperMCServer custom resource in Kubernetes.
        The server will be automatically reconciled by the operator.
      operationId: createServer
      tags:
        - servers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreateRequest"
            example:
              name: creative-server
              namespace: minecraft
              updateStrategy: auto
              updateDelay: "168h"
              checkCron: "0 3 * * *"
              maintenanceWindow:
                enabled: true
                cron: "0 4 * * 0"
              gracefulShutdownTimeout: "5m"
              rcon:
                enabled: true
                passwordSecretName: creative-rcon-secret
                passwordSecretKey: password
                port: 25575
      responses:
        "201":
          description: Server created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /plugins:
    get:
      summary: List all plugins
      description: |
        Returns a list of all Plugin resources in the cluster.
        Can be filtered by namespace using query parameter.
      operationId: listPlugins
      tags:
        - plugins
      parameters:
        - name: namespace
          in: query
          description: Filter plugins by Kubernetes namespace
          required: false
          schema:
            type: string
            example: minecraft
      responses:
        "200":
          description: Successfully retrieved list of plugins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginListResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new plugin
      description: |
        Creates a new Plugin custom resource in Kubernetes.
        The plugin will be automatically reconciled and matched to servers
        based on the instance selector.
      operationId: createPlugin
      tags:
        - plugins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginCreateRequest"
            example:
              name: bluemap
              namespace: minecraft
              source:
                type: hangar
                project: BlueMap
              updateStrategy: latest
              updateDelay: "168h"
              instanceSelector:
                matchLabels:
                  game-mode: survival
      responses:
        "201":
          description: Plugin created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Invalid request: missing required field 'name'"
            code: INVALID_REQUEST
            details:
              field: name
              reason: required
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "PaperMCServer 'my-server' already exists in namespace 'minecraft'"
            code: ALREADY_EXISTS
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Failed to communicate with Kubernetes API"
            code: INTERNAL_ERROR
  schemas:
...
    UpdateStrategy:
      type: string
      description: |
        Defines how versions are managed for servers and plugins.

        - **latest**: Always use the latest available version from the repository
        - **auto**: Use constraint solver to find best version compatible with all dependencies
        - **pin**: Pin to specific version, auto-update to latest build within that version
        - **build-pin**: Pin to specific version AND build number, no automatic updates
      enum:
        - latest
        - auto
        - pin
        - build-pin
      example: auto
...
    ServerListResponse:
      type: object
      required:
        - servers
      properties:
        servers:
          type: array
          description: List of server summaries
          items:
            $ref: "#/components/schemas/ServerSummary"
    ServerSummary:
      type: object
      description: Summary information about a PaperMCServer
      required:
        - name
        - namespace
        - updateStrategy
        - status
      properties:
        name:
          type: string
          description: Kubernetes resource name
          minLength: 1
          maxLength: 253
          example: survival-server
        namespace:
          type: string
          description: Kubernetes namespace
          example: minecraft
        currentVersion:
          type: string
          description: Currently running Paper version with build number (e.g., "1.21.1-100")
          example: "1.21.1-100"
        desiredVersion:
          type: string
          description: Target Paper version the operator wants to run
          example: "1.21.1-105"
        updateStrategy:
          $ref: "#/components/schemas/UpdateStrategy"
        status:
          $ref: "#/components/schemas/ServerStatus"
        pluginCount:
          type: integer
          description: Number of plugins matched to this server
          minimum: 0
          example: 5
        availableUpdate:
          $ref: "#/components/schemas/AvailableUpdate"
        nextMaintenance:
          type: string
          description: Human-readable description of next maintenance window
          example: "Every Sunday at 4:00"
        labels:
          type: object
          description: Kubernetes labels from the server resource
          additionalProperties:
            type: string
          example:
            environment: production
            game-mode: survival
    ServerDetail:
      type: object
      description: Detailed information about a PaperMCServer including plugins and schedules
      allOf:
        - $ref: "#/components/schemas/ServerSummary"
        - type: object
          properties:
            currentBuild:
              type: integer
              description: Currently running Paper build number
              minimum: 1
              example: 100
            desiredBuild:
              type: integer
              description: Target Paper build number
              minimum: 1
              example: 105
            plugins:
              type: array
              description: List of plugins matched to this server
              items:
                $ref: "#/components/schemas/ServerPlugin"
            maintenanceSchedule:
              $ref: "#/components/schemas/MaintenanceSchedule"
            lastUpdate:
              $ref: "#/components/schemas/UpdateHistory"
            updateBlocked:
              $ref: "#/components/schemas/UpdateBlockedStatus"
            conditions:
              type: array
              description: Kubernetes conditions for the server
              items:
                $ref: "#/components/schemas/Condition"
            rcon:
              $ref: "#/components/schemas/RCONStatus"
            service:
              $ref: "#/components/schemas/ServiceStatus"
...
    ServerCreateRequest:
      type: object
      description: Request body for creating a new PaperMCServer
      required:
        - name
        - namespace
        - updateStrategy
      properties:
        name:
          type: string
          description: Kubernetes resource name (must be valid DNS subdomain)
          minLength: 1
          maxLength: 253
          pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
          example: creative-server
        namespace:
          type: string
          description: Kubernetes namespace to create the server in
          example: minecraft
        updateStrategy:
          $ref: "#/components/schemas/UpdateStrategy"
        version:
          type: string
          description: |
            Target Paper version (required for pin and build-pin strategies).
            Example: "1.21.1" for Minecraft 1.21.1
          pattern: "^\\d+\\.\\d+(\\.\\d+)?$"
          example: "1.21.1"
        build:
          type: integer
          description: |
            Target Paper build number (required for build-pin strategy).
            For pin strategy, serves as minimum build.
          minimum: 1
          example: 100
        updateDelay:
          type: string
          description: |
            Grace period before applying updates (Go duration format).
            Examples: "168h" (7 days), "24h", "30m"
          pattern: "^(\\d+h)?(\\d+m)?(\\d+s)?$"
          example: "168h"
        checkCron:
          type: string
          description: |
            Cron expression for update checks.
            Format: minute hour day month weekday
          example: "0 3 * * *"
        maintenanceWindow:
          $ref: "#/components/schemas/MaintenanceWindowConfig"
        gracefulShutdownTimeout:
          type: string
          description: |
            Graceful shutdown timeout (Go duration format).
            Must match StatefulSet terminationGracePeriodSeconds.
          default: "5m"
          example: "5m"
        rcon:
          $ref: "#/components/schemas/RCONConfig"
        service:
          $ref: "#/components/schemas/ServiceConfig"
        labels:
          type: object
          description: Labels to apply to the server resource
          additionalProperties:
            type: string
          example:
            environment: production
            game-mode: creative
...
    PluginListResponse:
      type: object
      required:
        - plugins
      properties:
        plugins:
          type: array
          description: List of plugin summaries
          items:
            $ref: "#/components/schemas/PluginSummary"
...
    PluginDetail:
      type: object
      description: Detailed information about a Plugin
      allOf:
        - $ref: "#/components/schemas/PluginSummary"
        - type: object
          properties:
            updateDelay:
              type: string
              description: Grace period before applying updates
              example: "168h"
            port:
              type: integer
              description: Network port exposed by the plugin (added to Service)
              minimum: 1
              maximum: 65535
              example: 8123
            lastFetched:
              type: string
              format: date-time
              description: When metadata was last fetched from repository
              example: "2025-01-15T10:30:00Z"
            availableVersions:
              type: array
              description: Cached version metadata from repository
              items:
                $ref: "#/components/schemas/PluginVersion"
            matchedInstances:
              type: array
              description: List of servers this plugin is matched to
              items:
                $ref: "#/components/schemas/MatchedInstance"
            instanceSelector:
              $ref: "#/components/schemas/LabelSelector"
            compatibilityOverride:
              $ref: "#/components/schemas/CompatibilityOverride"
            deletionProgress:
              type: array
              description: JAR cleanup progress during plugin deletion
              items:
                $ref: "#/components/schemas/DeletionProgressEntry"
            conditions:
              type: array
              description: Kubernetes conditions for the plugin
              items:
                $ref: "#/components/schemas/Condition"
...
    PluginCreateRequest:
      type: object
      description: Request body for creating a new Plugin
      required:
        - name
        - namespace
        - source
        - instanceSelector
      properties:
        name:
          type: string
          description: Kubernetes resource name
          minLength: 1
          maxLength: 253
          pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
          example: bluemap
        namespace:
          type: string
          description: Kubernetes namespace
          example: minecraft
        source:
          $ref: "#/components/schemas/PluginSource"
        updateStrategy:
          $ref: "#/components/schemas/UpdateStrategy"
        version:
          type: string
          description: Pinned version (required for pin/build-pin strategies)
          example: "5.4"
        build:
          type: integer
          description: Pinned build number (for build-pin strategy)
          minimum: 1
        updateDelay:
          type: string
          description: Grace period before applying updates
          example: "168h"
        port:
          type: integer
          description: Network port to expose via Service
          minimum: 1
          maximum: 65535
          example: 8123
        instanceSelector:
          $ref: "#/components/schemas/LabelSelector"
        compatibilityOverride:
          $ref: "#/components/schemas/CompatibilityOverride"
...
    MaintenanceWindowConfig:
      type: object
      description: Maintenance window configuration for create/update requests
      properties:
        enabled:
          type: boolean
          description: Whether to enable maintenance window
          default: false
          example: true
        cron:
          type: string
          description: Cron expression for maintenance window
          example: "0 4 * * 0"
...
    ErrorResponse:
      type: object
      description: Error response body
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Failed to get server: not found"
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - NOT_FOUND
            - ALREADY_EXISTS
            - INTERNAL_ERROR
            - FORBIDDEN
            - UNAUTHORIZED
          example: NOT_FOUND
        details:
          type: object
          description: Additional error details
          additionalProperties: true
          example:
            field: name
            reason: invalid format